- merge cosmetic changes in batman, add define for header inclusion in hash.h
git-svn-id: http://downloads.open-mesh.org/svn/batman/trunk/batman@691 45894c77-fb22-0410-b583-ff6e7d5dbf6c

+++ bitarray.c
@@ -31,7 +31,7 @@ void bit_init( TYPE_OF_WORD *seq_bits ) {
 
 	int i;
 
-	for (i=0; i< NUM_WORDS; i++)
+	for (i = 0 ; i < (int)NUM_WORDS; i++)
 		seq_bits[i]= 0;
 
 };
